펀딩 게시물 등록 시 파일 업로드 다운로드 구현

+++ idolsns/src/main/java/com/kh/idolsns/dto/FundDto.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @Data
 public class FundDto {
 	
-	private int fundNo;
-	private int postNo;
+	private Long fundNo;
+	private Long postNo;
 	private int fundPrice;
 	private String fundId;
 	private Date fundTime;
